break up lists
