+++ src/python/zigopt/common/__init__.py
@@ -1,9 +1,15 @@
 # Copyright © 2022 Intel Corporation
 #
 # SPDX-License-Identifier: Apache License 2.0
+from zigopt.common.compact import *
+from zigopt.common.find import *
 from zigopt.common.functions import *
+from zigopt.common.generators import *
 from zigopt.common.lists import *
 from zigopt.common.non_crypto_random import *
 from zigopt.common.numbers import *
 from zigopt.common.objects import *
+from zigopt.common.recursive import *
+from zigopt.common.remove_nones import *
 from zigopt.common.strings import *
+from zigopt.common.types import *
