+++ src/python/zigopt/common/sigopt_datetime.py
@@ -4,13 +4,26 @@
 # crosshair: on
 import datetime
 import os
-from typing import Optional as _Optional
+from typing import Optional
 
 import deal
 import pytz
 
 
 # sigoptlint: disable=AvoidDatetimeNowRule
+__all__ = [
+  "default_timezone",
+  "datetime_to_seconds",
+  "seconds_to_datetime",
+  "unix_epoch",
+  "aware_datetime_to_naive_datetime",
+  "naive_datetime_to_aware_datetime",
+  "current_datetime",
+  "unix_timestamp",
+  "unix_timestamp_with_microseconds",
+  "get_month_interval",
+  "seconds_until_next_interval",
+]
 
 
 def default_timezone() -> datetime.tzinfo:
@@ -76,7 +89,7 @@ def _get_month_interval(dt: datetime.datetime) -> tuple[datetime.datetime, datet
   return start, end
 
 
-def get_month_interval(dt: _Optional[datetime.datetime] = None) -> tuple[datetime.datetime, datetime.datetime]:
+def get_month_interval(dt: Optional[datetime.datetime] = None) -> tuple[datetime.datetime, datetime.datetime]:
   if dt is None:
     dt = current_datetime()
   return _get_month_interval(dt)
